调整初始化数据，完善代码集功能

+++ HealthTeam/src/main/java/com/ht/clinic/dao/HealthCodeDefineDao.java
@@ -17,7 +17,7 @@ import org.apache.ibatis.annotations.Mapper;
 @Mapper
 public interface HealthCodeDefineDao {
 
-	HealthCodeDefineDO get(Long id);
+	HealthCodeDefineDO get(Integer id);
 	
 	List<HealthCodeDefineDO> list(Map<String, Object> map);
